image_backend/requirements.txt
Django==2.1
djangorestframework
psycopg2